Fixed the security issues (Week8)

+++ Week7/demo/src/main/resources/dbMigration/V1_0__my_first_migration.sql
@@ -1,6 +1,6 @@
 create table сpu(
 id integer auto_increment primary key not null,
 model varchar(50) not null,
-nucleus integer not null,
+nucleus smallint not null,
 frequency double not null
 );
